refactor: move game into a features folder
