+++ src/features/game/components/game-text/game-text.tsx
