+++ src/features/game/components/game-text/index.ts
