+++ src/features/game/components/user-info-block/index.ts
