+++ src/features/game/components/user-info-block/user-info-block.tsx
