+++ src/features/game/constants/user-info.constants.ts
